2x image rendering at 1x on iPod Touch with iOS 7.1.2
While testing on a 5th gen iPod Touch with iOS 7.1.2, it seems that specifying the "@2x" identifier effectively renders the @2x launch image at 1x. (Debugging the returned image's `.scale` property prints "1.0".) Commenting this line out and allowing iOS to presence check (and use) the @2x identifier seems to fix the issue. (Verified on a retina iPad with iOS 8.0 an an iPhone 6 with iOS 8.1.1)